issue/#68 move all files to the same path that related with NavBar
